% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphoci.R
\name{aphoci}
\alias{aphoci}
\title{Family of confidence interval routines}
\description{
aphoci style confidence intervals and tests
made easy with R
}
\details{
Output Functions

aphoci_rate(num, den, mult, level)
aphoci_prop(num, den, mult, level)

testci(ci.ref, ci.var, bSenseHigherisBetter, iVerbose)
testci_s(ci.ref, ci.var, bSenseHigherisBetter, iVerbose)
}
\seealso{
Other confidence interval routines: \code{\link{aphoci_gen}},
  \code{\link{calcci}}, \code{\link{stat_test}},
  \code{\link{testci}}, \code{\link{testspc}}
}
\concept{confidence interval routines}
