% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcci.R
\name{calcci}
\alias{calcci}
\title{Family of calculate routines}
\description{
Container for calculate routines
}
\details{
Routines to calculate confidence intervals.

Implementation details

APHO CI routines - but in R

Need to protect against some corner cases

prop.test

(x, 0) (NA, x) (x, NA) generates an error (0,  >0..<10) generates a warning

poisson.test

(0, 0) ci : (NaN, Inf) (x, 0) ci : (Inf, Inf) (NA, x) (x, NA) generates an
error

binom.test ??

For vectors, need to assemble as 1D list of pairs of (num, den, mult,
ci_type) return type compatible with data.frame.  Use "t" prefix for
data.table use.
}
\seealso{
Other calculate routines: \code{\link{aphoci_gen}}

Other confidence interval routines: \code{\link{aphoci_gen}},
  \code{\link{aphoci}}, \code{\link{stat_test}},
  \code{\link{testci}}, \code{\link{testspc}}
}
\concept{calculate routines}
\concept{confidence interval routines}
