% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{describe}
\alias{describe}
\title{My describe}
\usage{
describe(x, width = NULL, showvalues = FALSE, type = c("describe",
  "tabular"))
}
\arguments{
\item{x}{(data.frame) data to gimpse}

\item{width}{(int) console width, NULL means use default options()$width}

\item{showvalues}{(bool) Show field values (default TRUE)}

\item{type}{(character) specify describe or tabular lists.}
}
\value{
\preformatted{
#' \describe\{
#'   \item\{\code\{field1\}\}\{description 1\}
#'   \item\{\code\{field2\}\}\{description 2\}
#'  ...
#' \}

#' \tabular\{ll\}\{
#'   \{\{\code\{field1\}\} \tab \{description 1\} \cr
#'   \{\{\code\{field2\}\} \tab \{description 1\} \cr
#'   ...
#' \}
}
}
\description{
Helper for generating roxygen documentation for a data.frame
}
