% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_all}
\alias{extract_all}
\title{Extract all qof data}
\usage{
extract_all(qof_root = c("qof-1617", "qof-1516", "qof-1415", "qof-1314",
  "qof-1213", "qof-1112", "qof-1011", "qof-0910", "qof-0809", "qof-0708",
  "qof-0607", "qof-0506", "qof-0405"), bSaveData = FALSE)
}
\arguments{
\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
\code{qof-YYZZ}}

\item{bSaveData}{(boolean) Flag to determine whether to save datasets as
\code{devtools::use_data}}
}
\description{
Extract all qof data
}
