% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg_91_qof.R
\name{f__91__amend_data__add_subtotals}
\alias{f__91__amend_data__add_subtotals}
\title{Amend data by adding subtotals}
\usage{
f__91__amend_data__add_subtotals(qof, bCalcEngTotal = FALSE,
  bCalcCCGTotals = FALSE, lu.orgs.ccgs.local = NA,
  lu.orgs.ccgs.groups = NA)
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__91__load_raw}})}

\item{bCalcEngTotal}{Add an 'eng' that is total over all practices.}

\item{bCalcCCGTotals}{Add CCG totals (group practices by ccg_code)}

\item{lu.orgs.ccgs.local}{Filter on these ccgs (ccg_code)}

\item{lu.orgs.ccgs.groups}{Groups of ccgs (ccg_code, practice_code -> type, instance)}
}
\value{
list of lists (see \code{\link{f__91__load_raw}})
}
\description{
Can add england totals, ccg totals, filter for local ccgs, and group by given lookup.
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}
}
