% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__91__amend_orgref__ccg_groups}
\alias{f__91__amend_orgref__ccg_groups}
\title{Merge ccg groups into orgref}
\usage{
f__91__amend_orgref__ccg_groups(qof, lu.orgs.ccgs.groups = NA)
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__91__load_raw}})}

\item{lu.orgs.ccgs.groups}{ccg group lookup (see \code{\link{main}})}
}
\value{
list of lists (see \code{\link{f__91__load_raw}})
}
\description{
Merge ccg groups into orgref
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}

Other Process routines: \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__process__reference_measures_compare}}
}
\concept{Internal routines}
\concept{Process routines}
