% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__91__compare}
\alias{f__91__compare}
\title{Compare routines}
\usage{
f__91__compare(qof_measures, bWriteCSV = TRUE, qof_root,
  file_suffix = "__eng_ccg_prac__compare__bench_spc23__eng_ccg")
}
\arguments{
\item{qof_measures}{measures data frame}

\item{bWriteCSV}{Flag to indicate to write results to file.}

\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}

\item{file_suffix}{For loading and saving of any processed data}
}
\value{
compare data frame
}
\description{
Benchmark against reference value e.g. England - CI overlap with reference
SPC methods - compare point value with control limits
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{f__extract__load_raw}},
  \code{\link{families}}

Other Compare routines: \code{\link{f__91__load_compare}},
  \code{\link{families}}
}
\concept{Compare routines}
\concept{Internal routines}
