% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__91__load_measures}
\alias{f__91__load_measures}
\title{Load measures}
\usage{
f__91__load_measures(qof_root,
  file_suffix = "__eng_ccg_prac__measure_ndv")
}
\arguments{
\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}

\item{file_suffix}{For loading and saving of any processed data}
}
\value{
measures data frame
}
\description{
Load measures
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}

Other Load routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{families}}

Other Measure routines: \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}}, \code{\link{families}}
}
\concept{Internal routines}
\concept{Load routines}
\concept{Measure routines}
