% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__91__load_reference}
\alias{f__91__load_reference}
\title{Load reference}
\usage{
f__91__load_reference(qof_root, file_suffix = "__processed")
}
\arguments{
\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}

\item{file_suffix}{For loading and saving of any processed data}
}
\value{
reference list with named items \itemize{\item{orgref}\item{indmap}}
}
\description{
Load reference
}
\seealso{
Other Internal routines: \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__save_reference}},
  \code{\link{f__extract__load_raw}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Load routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__extract__load_raw}},
  \code{\link{families}}

Other Reference routines: \code{\link{f__91__save_reference}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Load routines}
\concept{Reference routines}
