% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg_91_qof.R
\name{f__91__measures}
\alias{f__91__measures}
\title{Calculate QOF measures}
\usage{
f__91__measures(qof, bWriteCSV = FALSE, qof_root,
  file_suffix = "__eng_ccg_prac__measure_ndv")
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__91__load_raw}})}

\item{bWriteCSV}{Flag to indicate to write results to file.}

\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}

\item{file_suffix}{For loading and saving of any processed data}
}
\description{
Wrapper around measures_ind and measures_prev.
}
\details{
Separate at this stage as ind and prev although a  lot of similarity have
different underlying data structure.
}
\note{
Call tree:

f__91__measures_ind
f__91__measures_prev
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}

Other Process routines: \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__process__reference_measures_compare}}

Other Measure routines: \code{\link{f__91__load_measures}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{families}}
}
