% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__91__preprocess}
\alias{f__91__preprocess}
\title{Preprocess}
\usage{
f__91__preprocess(qof, bWriteCSV = FALSE)
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__91__load_raw}})}

\item{bWriteCSV}{Flag to indicate to write results to file.}
}
\value{
a list of lists with named items
\describe{
    \item{reference}{
        \itemize{\item{orgref}\item{indmap}}
    }
    \item{data}{
        \itemize{\item{prev}\item{ind}\item{prev.melt}}
    }
}
}
\description{
optionally save tweaked reference data
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_raw}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}

Other Process routines: \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__process__reference_measures_compare}}
}
\concept{Internal routines}
\concept{Process routines}
