% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg_91_qof.R
\name{f__91__process__reference_measures_compare}
\alias{f__91__process__reference_measures_compare}
\title{Process all}
\usage{
f__91__process__reference_measures_compare(qof_period = c("1516", "1617"),
  lu.orgs.ccgs.local = c("02Q", paste0("04", c("E", "H", "K", "L", "M",
  "N"))), lu.orgs.ccgs.groups = NA, bWriteCSV = FALSE)
}
\arguments{
\item{qof_period}{Of the form "YYZZ"}

\item{lu.orgs.ccgs.local}{vector of strings for health codes e.g. c("04K", "04E")}

\item{lu.orgs.ccgs.groups}{lookup table (see #code{\link{main}})}

\item{bWriteCSV}{Flag to indicate to write results to file.}
}
\value{
a list with named items
\describe{
    \item{data}{Raw numbers}
    \item{reference}{Reference tables}
    \item{measures}{Measures}
    \item{compare}{Comparisons}
}
}
\description{
Load raw data and produce measures and compare against england.
}
\note{
Call tree:

f__91__load_data
f__91__amend_data__add_subtotals
f__91__amend_orgref__ccg_groups
f__91__save_reference
f__91__measures
f__91__compare
}
\seealso{
Other External routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{families}}

Other Process routines: \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}}
}
