% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__91__save_reference}
\alias{f__91__save_reference}
\title{Save reference}
\usage{
f__91__save_reference(qof, qof_root, file_suffix = "__processed",
  bWriteCSV = TRUE)
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__extract__load_raw}})}

\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}

\item{file_suffix}{For loading and saving of any processed data}

\item{bWriteCSV}{Flag to indicate to write results to file.}
}
\value{
Not sure, possibly NULL.
}
\description{
Save reference data
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__extract__load_raw}},
  \code{\link{families}}

Other Save routines: \code{\link{families}}

Other Reference routines: \code{\link{f__91__load_reference}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Reference routines}
\concept{Save routines}
