% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{f__combine_any}
\alias{f__combine_any}
\title{Merge all qof data into one dataset}
\usage{
f__combine_any(qof_root = c("qof-1718", "qof-1617", "qof-1516",
  "qof-1415", "qof-1314", "qof-1213", "qof-1112", "qof-1011", "qof-0910",
  "qof-0809", "qof-0708", "qof-0607", "qof-0506", "qof-0405"),
  bSaveData = FALSE)
}
\arguments{
\item{qof_root}{(character)

  Directory root for loading and saving any processed data.  Of the form
  \code{qof-YYZZ}}

\item{bSaveData}{(bool) Flag to determine whether to save datasets as
\code{devtools::use_data}}
}
\value{
a list with named items
\tabular{ll}{
  \code{meta_org}  \tab Organisation metadata \cr
  \code{meta_ind}  \tab Indicator metadata \cr
  \code{data_prev} \tab Prevalence data \cr
  \code{data_ind}  \tab Indicator data \cr
}
}
\description{
Merges data from multiple qof periods.  Uses the post-extract results.  Binds
qof periods into relevant data.frames.
}
