% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{f__extract}
\alias{f__extract}
\title{Return extracted dataset}
\usage{
f__extract(qof_root = c("qof-1718", "qof-1617", "qof-1516", "qof-1415",
  "qof-1314", "qof-1213", "qof-1112", "qof-1011", "qof-0910", "qof-0809",
  "qof-0708", "qof-0607", "qof-0506", "qof-0405"),
  bExtractFromRaw = FALSE)
}
\arguments{
\item{qof_root}{(character)

  Directory root for loading and saving any processed data.  Of the form
  \code{qof-YYZZ}}

\item{bExtractFromRaw}{(bool) Flag to do process raw or load pre-extracted.}
}
\description{
Either extract from raw values or load pre-extracted values.
}
