% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__main__load__reference_measures_compare}
\alias{f__main__load__reference_measures_compare}
\title{Load processed}
\usage{
f__main__load__reference_measures_compare(qof_period = c("1617", "1516"),
  bLoadData = FALSE)
}
\arguments{
\item{qof_period}{of the form "ZZYY"}

\item{bLoadData}{Specify to load raw numbers too.}
}
\value{
a list with named items
\describe{
    \item{data}{Raw numbers}
    \item{reference}{Reference tables}
    \item{measures}{Measures}
    \item{compare}{Comparisons}
}
}
\description{
Load pre-processed measures.  Optionally load the raw numbers behind the
measures.
}
\note{
Call tree:

f__main__load_reference
f__main__load_measures
f__main__load_compare
f__main__load_data
}
\seealso{
Other External routines: \code{\link{f__process__reference_measures_compare}},
  \code{\link{families}}

Other Load routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{families}}
}
\concept{External routines}
\concept{Load routines}
