% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__main__load_data}
\alias{f__main__load_data}
\title{load raw data}
\usage{
f__main__load_data(qof_root)
}
\value{
(list of data.frame objects) reference list with named items
  \itemize{\item{data_ind}\item{data_prev}}
}
\description{
Default is not to do anything further.
}
\note{
Call tree:

f__extract__load_raw
f__main__preprocess
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Load routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load__reference_measures_compare}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Load routines}
