% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__main__load_measures}
\alias{f__main__load_measures}
\title{Load measures}
\usage{
f__main__load_measures(qof_root,
  file_suffix = "__eng_ccg_prac__measure_ndv")
}
\arguments{
\item{qof_root}{(character)

  Directory root for loading and saving any processed data.  Of the form
  \code{qof-YYZZ}}

\item{file_suffix}{For loading and saving of any processed data}
}
\value{
(data.frame) measures data frame
}
\description{
Load measures
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_reference}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Load routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load__reference_measures_compare}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_reference}},
  \code{\link{families}}

Other Measure routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Load routines}
\concept{Measure routines}
