% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__main__load_reference}
\alias{f__main__load_reference}
\title{Load reference}
\usage{
f__main__load_reference(qof_root)
}
\value{
(list of data.frame objects) reference list with named items
  \itemize{\item{meta_org}\item{meta_ind}}
}
\description{
Load reference
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Load routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load__reference_measures_compare}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{families}}

Other Reference routines: \code{\link{f__main__save_reference}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Load routines}
\concept{Reference routines}
