% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__process__calc_measures_prev}
\alias{f__process__calc_measures_prev}
\title{Create QOF prevalance measures}
\usage{
f__process__calc_measures_prev(qof, bWriteCSV = FALSE,
  file_suffix = "__eng_ccg_prac__measure_ndv", bSaveData = TRUE)
}
\arguments{
\item{qof}{list of lists (see \code{\link{f__extract__load_raw}})}

\item{bWriteCSV}{Flag to indicate to write results to file.}

\item{file_suffix}{For loading and saving of any processed data}

\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
"qof-YYZZ"}
}
\value{
prevalence data frame
}
\description{
Create QOF prevalance measures
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Process routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__measures}},
  \code{\link{f__process__reference_measures_compare}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}}

Other Measure routines: \code{\link{f__main__load_measures}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__measures}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Measure routines}
\concept{Process routines}
