% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__process__compare}
\alias{f__process__compare}
\title{Compare routines}
\usage{
f__process__compare(qof_measures, bWriteCSV = FALSE,
  file_suffix = "__eng_ccg_prac__compare__bench_spc23__eng_ccg",
  bSaveData = TRUE)
}
\arguments{
\item{qof_measures}{(data.frame) measures data frame}

\item{bWriteCSV}{(bool) Flag to indicate to write results to file.}

\item{file_suffix}{(character) For loading and saving of any processed data}

\item{bSaveData}{(bool) Flag to store as package data}
}
\value{
(data.frame) compare data frame
}
\description{
Benchmark against reference value e.g. England - CI overlap with reference
SPC methods - compare point value with control limits
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Compare routines: \code{\link{f__main__load_compare}},
  \code{\link{families}}
}
\concept{Compare routines}
\concept{Internal routines}
