% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__process__measures}
\alias{f__process__measures}
\title{Calculate QOF measures}
\usage{
f__process__measures(qof, bWriteCSV = FALSE,
  file_suffix = "__eng_ccg_prac__measure_ndv", bSaveData = TRUE)
}
\arguments{
\item{qof}{(list of lists) (see \code{\link{f__extract__load_raw}})}

\item{bWriteCSV}{(bool) Flag to indicate to write results to file.}

\item{file_suffix}{(character) For loading and saving of any processed data}

\item{bSaveData}{(bool) Flag to store as package data}
}
\description{
Wrapper around measures_ind and measures_prev.
}
\details{
Separate at this stage as ind and prev although a  lot of similarity have
different underlying data structure.
}
\note{
Call tree:

f__main__measures_ind
f__main__measures_prev
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}},
  \code{\link{qof__main__load_compare}},
  \code{\link{qof__main__load_data}},
  \code{\link{qof__main__load_measures}},
  \code{\link{qof__main__load_reference}}

Other Process routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__reference_measures_compare}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}}

Other Measure routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{families}},
  \code{\link{qof__main__load_measures}}
}
\concept{Internal routines}
\concept{Measure routines}
\concept{Process routines}
