% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{f__process__reference_measures_compare}
\alias{f__process__reference_measures_compare}
\title{Process all}
\usage{
f__process__reference_measures_compare(qof, bWriteCSV = FALSE)
}
\arguments{
\item{qof}{(list of data.frames) result of \code{f__transform}}

\item{bWriteCSV}{Flag to indicate to write results to file.}
}
\value{
a list with named items
\describe{
    \item{qof (list)}{A list of data_ind, data_prev, meta_ind, meta_prev tables}
    \item{measures (data.frame)}{Measures}
    \item{compare (dtat.frame)}{Comparisons}
}
}
\description{
Load raw data and produce measures and compare against england.
}
\note{
Call tree:

f__91__load_data
f__91__amend_data__add_subtotals
f__91__amend_orgref__ccg_groups
f__91__save_reference
f__91__measures
f__91__compare
}
\seealso{
f__transform

Other External routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{families}}

Other Process routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}}
}
\concept{External routines}
\concept{Process routines}
