% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform}
\alias{f__transform}
\title{Apply all transforms}
\usage{
f__transform(qof, lu_ccgs = NA, lu_ccg_groups = NA)
}
\arguments{
\item{qof}{list of data items (see \code{\link{f__extract__load_raw}})}

\item{lu_ccg_groups}{local ccg group lookups.  NA means use Notts. default.}
}
\description{
Apply all transforms
}
\seealso{
f__transform__create_local_lu
}
