% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform__create_local_lu}
\alias{f__transform__create_local_lu}
\title{Default local lookups}
\usage{
f__transform__create_local_lu(bSaveData = FALSE)
}
\arguments{
\item{bSaveData}{flag to write use_data sets}
}
\value{
list of lu_ccgs, lu_ccg_groups
}
\description{
Default local lookups
}
