% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform__data__add_orgtype}
\alias{f__transform__data__add_orgtype}
\title{Add org.type field to data}
\usage{
f__transform__data__add_orgtype(qof)
}
\arguments{
\item{qof}{list of data items (see \code{\link{f__extract__load_raw}})}
}
\value{
list of tables
}
\description{
Modifies dat_ind and data_prev tables by adding a field \code{org.type} with
value \code{ccg::practice}
}
