% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform__data__add_subtotals}
\alias{f__transform__data__add_subtotals}
\title{Amend data by adding subtotals}
\usage{
f__transform__data__add_subtotals(qof, bCalcEngTotal = FALSE,
  bCalcCCGTotals = FALSE, lu_ccgs = NA, lu_ccg_groups = NA)
}
\arguments{
\item{qof}{list of data items (see \code{\link{f__extract__load_raw}})}

\item{bCalcEngTotal}{Add an 'eng' that is total over all practices.}

\item{bCalcCCGTotals}{Add CCG totals (group practices by ccg_code)}

\item{lu_ccgs}{Filter on these ccgs (ccg_code)}

\item{lu_ccg_groups}{Groups of ccgs (ccg_code, practice_code -> type,
instance)}
}
\value{
a list with named items
\tabular{ll}{
  \code{meta_org}  \tab Organisation metadata \cr
  \code{meta_ind}  \tab Indicator metadata \cr
  \code{data_prev} \tab Prevalence data \cr
  \code{data_ind}  \tab Indicator data  \cr
}
}
\description{
Can add england totals, ccg totals, filter for local ccgs, and group by given
lookup.
}
\seealso{
Other Internal routines: \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__save_reference}},
  \code{\link{f__extract__load_raw}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}
}
\concept{Internal routines}
\concept{Process routines


Consistently use practice_code as instance, ccg_code as type ... even for
CCGs ...}
