% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform__meta__ccg_groups}
\alias{f__transform__meta__ccg_groups}
\title{Merge ccg groups into orgref}
\usage{
f__transform__meta__ccg_groups(qof, lu_ccg_groups = NA)
}
\arguments{
\item{qof}{list of data items (see \code{\link{f__extract__load_raw}})}

\item{lu_ccg_groups}{ccg group lookup (see \code{\link{main}})}
}
\value{
a list with named items
\tabular{ll}{
  \code{meta_org}  \tab Organisation metadata \cr
  \code{meta_ind}  \tab Indicator metadata \cr
  \code{data_prev} \tab Prevalence data (molten on measure/value) \cr
  \code{data_ind}  \tab Indicator data  (molten on measure/value) \cr
}
}
\description{
Merge ccg groups into orgref
}
\seealso{
f__transform__create_local_lu

Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load_compare}},
  \code{\link{f__main__load_data}},
  \code{\link{f__main__load_measures}},
  \code{\link{f__main__load_reference}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}}

Other Process routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}},
  \code{\link{f__process__reference_measures_compare}},
  \code{\link{f__transform__preprocess}}
}
\concept{Internal routines}
\concept{Process routines}
