% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{f__transform__preprocess}
\alias{f__transform__preprocess}
\title{Preprocess}
\usage{
f__transform__preprocess(qof)
}
\arguments{
\item{qof}{list of data items (see \code{\link{f__extract__load_raw}})}
}
\value{
a list with named items
\tabular{ll}{
  \code{meta_org}  \tab Organisation metadata \cr
  \code{meta_ind}  \tab Indicator metadata \cr
  \code{data_prev} \tab Prevalence data (molten on measure/value) \cr
  \code{data_ind}  \tab Indicator data  (molten on measure/value) \cr
}
}
\description{
- melt data_ind, data_prev onto measure/value pairs
- tag on meta_org ccg and indicator_group, domain_code
- identify register indicators in meta_ind
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{families}},
  \code{\link{qof__main__load_compare}},
  \code{\link{qof__main__load_data}},
  \code{\link{qof__main__load_measures}},
  \code{\link{qof__main__load_reference}}

Other Process routines: \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}},
  \code{\link{f__process__reference_measures_compare}},
  \code{\link{f__transform__meta__ccg_groups}}
}
\concept{Internal routines}
\concept{Process routines}
