% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qofvariation.ncc-package.R
\name{families}
\alias{families}
\alias{family_of_helpers}
\alias{family_of_external}
\alias{family_of_internal}
\alias{family_of_load}
\alias{family_of_save}
\alias{family_of_measure}
\alias{family_of_compare}
\alias{family_of_reference}
\title{Families}
\description{
List of families.
}
\seealso{
Other Helper routines: \code{\link{is_installed}},
  \code{\link{proj_path}}, \code{\link{proj_root}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}

Other External routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__process__reference_measures_compare}}

Other Internal routines: \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__save_reference}},
  \code{\link{f__extract__load_raw}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}}

Other Load routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__extract__load_raw}}

Other Save routines: \code{\link{f__91__save_reference}}

Other Reference routines: \code{\link{f__91__load_reference}},
  \code{\link{f__91__save_reference}}

Other Measure routines: \code{\link{f__91__load_measures}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}}

Other Compare routines: \code{\link{f__91__load_compare}},
  \code{\link{f__process__compare}}

Other Helper routines: \code{\link{is_installed}},
  \code{\link{proj_path}}, \code{\link{proj_root}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}

Other External routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__process__reference_measures_compare}}

Other Internal routines: \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__save_reference}},
  \code{\link{f__extract__load_raw}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}}

Other Load routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__extract__load_raw}}

Other Save routines: \code{\link{f__91__save_reference}}

Other Measure routines: \code{\link{f__91__load_measures}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__measures}}

Other Compare routines: \code{\link{f__91__load_compare}},
  \code{\link{f__process__compare}}

Other Reference routines: \code{\link{f__91__load_reference}},
  \code{\link{f__91__save_reference}}
}
\concept{Compare routines}
\concept{External routines}
\concept{Helper routines}
\concept{Internal routines}
\concept{Load routines}
\concept{Measure routines}
\concept{Reference routines}
\concept{Save routines}
