% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fglimpse}
\alias{fglimpse}
\title{My glimpse}
\usage{
fglimpse(x, width = NULL, ...)
}
\arguments{
\item{x}{(data.frame) data to gimpse}

\item{width}{console width, NULL means use default options()$width}

\item{...}{(ignored) extra options

factor version of glimpse}
}
\description{
My glimpse
}
