% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fsummary}
\alias{fsummary}
\title{My summary}
\usage{
fsummary(x, ...)
}
\arguments{
\item{x}{data to summary-ise}

\item{...}{extra options to pass to summary

factor version of summary}
}
\description{
My summary
}
