% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilike}
\alias{ilike}
\alias{\%ilike\%}
\title{infix like but ignore case}
\usage{
ilike(vector, pattern)

vector \%ilike\% pattern
}
\arguments{
\item{vector}{Either a \code{character} vector or a \code{factor}.  A
\code{factor} is faster.}

\item{pattern}{Passed on to \code{grepl}}
}
\description{
infix like but ignore case
}
\section{Functions}{
\itemize{
\item \code{\%ilike\%}: 
}}

