% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{is_installed}
\alias{is_installed}
\title{check if package is installed}
\usage{
is_installed(p)
}
\arguments{
\item{p}{name of package (character)}
}
\description{
check if package is installed
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}
}
\concept{Helper routines}
