% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{l_readxl_wb}
\alias{l_readxl_wb}
\title{local worker to loop through list of xls}
\usage{
l_readxl_wb(this.file, f_readxl_ws = NULL, skipsheets = NULL, ...)
}
\arguments{
\item{this.file}{pathname of xl workbook}

\item{f_readxl_ws}{function to parse the workbook}

\item{skipsheets}{list of sheet anems to skip}

\item{...}{additional arguments passed to f_readxl_ws}
}
\value{
(data.frame likely) results of parsing workbook with
  \code{f_readxl_ws}
}
\description{
local worker to loop through list of xls
}
\seealso{
Other Extract routines: \code{\link{f__extract__load_raw}},
  \code{\link{l_readxl_ws_ind}},
  \code{\link{l_readxl_ws_prev}}
}
\concept{Extract routines}
