% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{l_readxl_ws_ind}
\alias{l_readxl_ws_ind}
\title{local worker to read through worksheets in workbook}
\usage{
l_readxl_ws_ind(this.sheet, this.file)
}
\arguments{
\item{this.sheet}{worksheet to parse the workbook}

\item{this.file}{pathname of xl workbook}
}
\value{
(data.frame likely) result of parsing the worksheet
}
\description{
variation for ind
}
\seealso{
Other Extract routines: \code{\link{f__extract__load_raw}},
  \code{\link{l_readxl_wb}}, \code{\link{l_readxl_ws_prev}}
}
\concept{Extract routines}
