% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_ccg_groups}
\alias{lu_ccg_groups}
\title{Local CCG groups}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 27 rows and 6 columns.}
\usage{
lu_ccg_groups
}
\description{
\tabular{ll}{
 \code{ccg_group_type} \tab {<chr n=4>} \cr
 \code{ccg_group_type_name} \tab {<chr n=4>} \cr
 \code{ccg_code} \tab {<chr n=7>} \cr
 \code{ccg_group_code} \tab {<chr n=7>} \cr
 \code{ccg_group_name} \tab {<chr n=7>} \cr
 \code{type_display_order} \tab {<int>} \cr
}
}
\keyword{datasets}
