% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_ccgs}
\alias{lu_ccgs}
\title{List of CCGs considered local}
\format{An object of class \code{character} of length 7.}
\usage{
lu_ccgs
}
\description{
List of CCGs considered local
}
\keyword{datasets}
