% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Process raw datasets}
\usage{
main(qof_period = c("1617", "1516"), bProcessRaw = FALSE,
  bWriteCSV = FALSE, bLoadData = FALSE)
}
\arguments{
\item{qof_period}{Of the form "YYZZ"}

\item{bProcessRaw}{if FALSE just load.  If TRUE process and then load.}

\item{bWriteCSV}{Flag to indicate to write results to file.}

\item{bLoadData}{Specify to load raw numbers too.}
}
\description{
Options to specify year. Option to specify if to run the raw process routines
Option to specify if to save results to disc as .csv Option to specify just
to load the data
}
\note{
Indirectly can re-create groups lookup (\code{lu__ccgs_groups.csv})

Process raw returns measures,reference data and raw data. Otherwise just
loads measures and reference data.  To additionally load raw data specify
\code{bLoadData = TRUE}
}
\examples{
\dontrun{
v1 <- main(qof_period = "1516", bProcessRaw = TRUE, bWriteCSV = TRUE)
v2 <- main(qof_period = "1516", bProcessRaw = FALSE, bLoadData = FALSE)
}
}
