% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{merge_all}
\alias{merge_all}
\title{Merge all qof data into one dataset}
\usage{
merge_all(qof_root = c("qof-1718", "qof-1617", "qof-1516", "qof-1415",
  "qof-1314", "qof-1213", "qof-1112", "qof-1011", "qof-0910", "qof-0809",
  "qof-0708", "qof-0607", "qof-0506", "qof-0405"), bSaveData = FALSE)
}
\arguments{
\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
\code{qof-YYZZ}}

\item{bSaveData}{(boolean) Flag to determine whether to save datasets as
\code{devtools::use_data}}
}
\description{
Essentially add qof_poeriod field to all tables
}
\examples{
load("./data/qof_meta_org.Rda")
load("./data/qof_meta_ind.Rda")
load("./data/qof_data_ind.Rda")
load("./data/qof_data_prev.Rda")

}
