% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{paste_paths}
\alias{paste_paths}
\title{Shortcut for constructing paths}
\usage{
paste_paths(...)
}
\arguments{
\item{...}{path elemnents.  Passed to paste.}
}
\description{
Shortcut for constructing paths
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{proj_root}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}
}
