% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcci.R
\name{poisson_byars.test}
\alias{poisson_byars.test}
\title{Byars approximation to poission}
\usage{
poisson_byars.test(x, T, conf.level = 0.95)
}
\arguments{
\item{x}{Numerator / number of events}

\item{T}{Denominator / time base for event count}

\item{conf.level}{Confidence level (\code{0 < level <= 1}).  Optional -
default 0.95}
}
\value{
Double vector dim 2 containing lower and upper confidence level.
}
\description{
Byars approximation to poission
}
\seealso{
Other statistic test routines: \code{\link{prop_wilson.test}},
  \code{\link{stat_test}}
}
\concept{statistic test routines}
