% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_path}
\alias{proj_path}
\title{Shortcut for constructing paths}
\usage{
proj_path(..., ext = "")
}
\arguments{
\item{...}{path elements.  Passed to paste.}
}
\description{
Shortcut for constructing paths
}
\note{
Written before usethis::proj_path was known ... very similar
  functionality so use usethis instead.
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{proj_root}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}
}
\concept{Helper routines}
