% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_root}
\alias{proj_root}
\title{Concept of 'root directory' depends on context.}
\usage{
proj_root()
}
\description{
- from project root is ./R
- from ./inst/dashboard it is, not surprisingly, ./R ...
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{proj_path}},
  \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}, \code{\link{transpose}}
}
\concept{Helper routines}
