% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcci.R
\name{prop_wilson.test}
\alias{prop_wilson.test}
\title{Wilson approximation to proportion}
\usage{
prop_wilson.test(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{Numerator / count of successes}

\item{n}{Denominator / number of trials}

\item{conf.level}{Confidence level (\code{0 < level <= 1}).  Optional -
default 0.95}
}
\value{
Double vector dim 2 containing lower and upper confidence level.
}
\description{
Wilson approximation to proportion
}
\seealso{
Other statistic test routines: \code{\link{poisson_byars.test}},
  \code{\link{stat_test}}
}
\concept{statistic test routines}
