% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_1617_data_ind}
\alias{qof_1617_data_ind}
\title{QOF data for indicators}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1882829 rows and 4 columns.}
\usage{
qof_1617_data_ind
}
\description{
\preformatted{
$ practice_code : chr
$ indicator_code: chr
$ measure       : chr
$ value         : num
}
}
\seealso{
Other qof_data: \code{\link{qof_1617_data_prev}}
}
\concept{qof_data}
\keyword{datasets}
