% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_1617_meta_ind}
\alias{qof_1617_meta_ind}
\title{QOF reference data for organisations}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 77 rows and 8 columns.}
\usage{
qof_1617_meta_ind
}
\description{
\preformatted{
$ indicator_code             : chr
$ indicator_description      : chr
$ indicator_point_value      : int
$ indicator_group_code       : chr
$ indicator_group_description: chr
$ domain_code                : chr
$ domain_description         : chr
$ patient_list_type          : chr
}
}
\seealso{
Other qof_meta: \code{\link{qof_1617_meta_org}}
}
\concept{qof_meta}
\keyword{datasets}
