% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qof__main}
\alias{qof__main}
\title{Process raw datasets}
\usage{
qof__main(qof_root = c("qof-1718", "qof-1617", "qof-1516", "qof-1415",
  "qof-1314", "qof-1213", "qof-1112", "qof-1011", "qof-0910", "qof-0809",
  "qof-0708", "qof-0607", "qof-0506", "qof-0405"),
  bExtractFromRaw = FALSE, bWriteCSV = FALSE, bSaveData = FALSE,
  bLoadData = FALSE)
}
\arguments{
\item{qof_root}{(character) Of the form "qof-YYZZ"}

\item{bExtractFromRaw}{(bool) if FALSE just load.  If TRUE process and then
load. Default FALSE.}

\item{bWriteCSV}{(bool) Flag to indicate to write results to file.  Default
FALSE.}

\item{bSaveData}{(bool) Specify to save measures and compare phases.  Default
FALSE.}

\item{bLoadData}{(bool) Specify to load raw numbers too.  Default FALSE.}
}
\description{
Options to specify year. Option to specify if to run the raw process routines
Option to specify if to save results to disc as .csv Option to specify just
to load the data
}
\note{
Indirectly can re-create groups lookup (\code{lu__ccgs_groups.csv})

Process raw returns measures,reference data and raw data. Otherwise just
loads measures and reference data.  To additionally load raw data specify
\code{bLoadData = TRUE}
}
\examples{
\dontrun{
v1 <- qof__main(qof_root = "1516", bExtractFromRaw = TRUE)
v2 <- qof__main(qof_root = "1516", bExtractFromRaw = FALSE)
}

}
