% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qof__main__load}
\alias{qof__main__load}
\title{Load preprocessed data}
\usage{
qof__main__load(qof_root = NULL)
}
\arguments{
\item{qof_root}{(character) Of the form "qof-YYZZ"}
}
\value{
(list of objects) list(qof = list(meta_org, meta_ind, data_ind,
  data_prev), measures, compare)
}
\description{
Make package provided data available as one onbect.
}
