% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qof__main__load_ccglu}
\alias{qof__main__load_ccglu}
\title{access loacl CCG lu}
\usage{
qof__main__load_ccglu()
}
\value{
list(lu-ccgs (character vector), lu_ccg_groups (data.frame))
}
\description{
access loacl CCG lu
}
