% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qof__main__load_compare}
\alias{qof__main__load_compare}
\title{Load compare}
\usage{
qof__main__load_compare(qof_root = NULL,
  file_suffix = "__eng_ccg_prac__compare__bench_spc23__eng_ccg")
}
\arguments{
\item{qof_root}{(character)

  Directory root for loading and saving any processed data.  Of the form
  \code{qof-YYZZ}}

\item{file_suffix}{For loading and saving of any processed data}
}
\value{
compare data frame
}
\description{
Load compare
}
\seealso{
Other Internal routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__save_reference}},
  \code{\link{f__process__calc_measures_ind}},
  \code{\link{f__process__calc_measures_prev}},
  \code{\link{f__process__compare}},
  \code{\link{f__process__measures}},
  \code{\link{f__transform__data__add_subtotals}},
  \code{\link{f__transform__meta__ccg_groups}},
  \code{\link{f__transform__preprocess}},
  \code{\link{families}},
  \code{\link{qof__main__load_data}},
  \code{\link{qof__main__load_measures}},
  \code{\link{qof__main__load_reference}}

Other Load routines: \code{\link{f__extract__load_raw}},
  \code{\link{f__main__load__reference_measures_compare}},
  \code{\link{families}},
  \code{\link{qof__main__load_data}},
  \code{\link{qof__main__load_measures}},
  \code{\link{qof__main__load_reference}}

Other Compare routines: \code{\link{f__process__compare}},
  \code{\link{families}}
}
\concept{Compare routines}
\concept{Internal routines}
\concept{Load routines}
