% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_compare}
\alias{qof_compare}
\title{QOF measure comparisons}
\format{An object of class \code{data.frame} with 1189251 rows and 13 columns.}
\usage{
qof_compare
}
\description{
Measures compared to benchmarks
}
\keyword{datasets}
