% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{qof_data}
\alias{qof_data}
\title{QOF data}
\description{
QOF data
}
\seealso{
Other qof_datasets: \code{\link{qof_datasets}},
  \code{\link{qof_meta}}

Other qof_data: \code{\link{qof_data_ind}},
  \code{\link{qof_data_prev}}, \code{\link{qof_datasets}}
}
\concept{qof_data}
\concept{qof_datasets}
