% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_data_ind}
\alias{qof_data_ind}
\alias{qof_1617_data_ind}
\alias{qof_1516_data_ind}
\alias{qof_1415_data_ind}
\alias{qof_1314_data_ind}
\alias{qof_1213_data_ind}
\title{QOF data for indicators}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1882829 rows and 4 columns.}
\usage{
qof_1617_data_ind

qof_1516_data_ind

qof_1415_data_ind

qof_1314_data_ind

qof_1213_data_ind
}
\description{
\preformatted{
$ practice_code : chr
$ indicator_code: chr
$ measure       : chr
$ value         : num
}
}
\details{
\describe{
  \item{\code{measure}}{
  One of
  \code{ACHIEVED_POINTS}
  , \code{DENOMINATOR}
  , \code{EXCEPTIONS}
  , \code{NUMERATOR}
  , \code{REGISTER}
  }
}
}
\seealso{
Other qof_data: \code{\link{qof_data_prev}},
  \code{\link{qof_datasets}}, \code{\link{qof_data}}
}
\concept{qof_data}
\keyword{datasets}
