% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_data_prev}
\alias{qof_data_prev}
\alias{qof_1617_data_prev}
\alias{qof_1516_data_prev}
\alias{qof_1415_data_prev}
\alias{qof_1314_data_prev}
\alias{qof_1213_data_prev}
\title{QOF data for prevalence}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 155232 rows and 5 columns.}
\usage{
qof_1617_data_prev

qof_1516_data_prev

qof_1415_data_prev

qof_1314_data_prev

qof_1213_data_prev
}
\description{
\preformatted{
$ practice_code       : chr
$ indicator_group_code: chr
$ register            : int
$ patient_list_type   : chr
$ patient_list_size   : int
}
}
\seealso{
Other qof_data: \code{\link{qof_data_ind}},
  \code{\link{qof_datasets}}, \code{\link{qof_data}}
}
\concept{qof_data}
\keyword{datasets}
