% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{qof_datasets}
\alias{qof_datasets}
\title{QOF datasets}
\description{
QOF datasets
}
\seealso{
Other qof_datasets: \code{\link{qof_data}},
  \code{\link{qof_meta}}

Other qof_data: \code{\link{qof_data_ind}},
  \code{\link{qof_data_prev}}, \code{\link{qof_data}}

Other qof_meta: \code{\link{qof_meta_ind}},
  \code{\link{qof_meta_org}}, \code{\link{qof_meta}}
}
\concept{qof_data}
\concept{qof_datasets}
\concept{qof_meta}
