% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_measures}
\alias{qof_measures}
\title{QOF measures}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 139416 rows and 12 columns.}
\usage{
qof_measures
}
\description{
indicator and prevalence as measures
}
\keyword{datasets}
