% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{qof_meta}
\alias{qof_meta}
\title{QOF meta}
\description{
QOF meta
}
\seealso{
Other qof_datasets: \code{\link{qof_datasets}},
  \code{\link{qof_data}}

Other qof_meta: \code{\link{qof_datasets}},
  \code{\link{qof_meta_ind}}, \code{\link{qof_meta_org}}
}
\concept{qof_datasets}
\concept{qof_meta}
