% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_meta_ind}
\alias{qof_meta_ind}
\title{QOF reference data for indicators}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 625 rows and 9 columns.}
\usage{
qof_meta_ind
}
\description{
\tabular{ll}{
 \code{indicator_code} \tab {<chr n=255>} \cr
 \code{indicator_description} \tab {<chr n=270>} \cr
 \code{indicator_point_value} \tab {<int>} \cr
 \code{indicator_group_code} \tab {<chr n=38>} \cr
 \code{indicator_group_description} \tab {<chr n=43>} \cr
 \code{domain_code} \tab {<chr n=6>} \cr
 \code{domain_description} \tab {<chr n=7>} \cr
 \code{patient_list_type} \tab {<chr n=11>} \cr
 \code{qof_period} \tab {<chr n=7>} \cr
}
}
\details{
\describe{
  \item{\code{patient_list_type}}{
  One of
  \code{TOTAL}
  , \code{15OV}, \code{16OV}, \code{17OV}, \code{18OV}
  , \code{45OV}, \code{50OV}
  , \code{25_64_F}
  , \code{30_74}
  , \code{54UN_F}
  }
}
}
\seealso{
Other qof_meta: \code{\link{qof_datasets}},
  \code{\link{qof_meta_org}}, \code{\link{qof_meta}}
}
\concept{qof_meta}
\keyword{datasets}
