% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_meta_org}
\alias{qof_meta_org}
\title{QOF reference data for organisations}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 22111 rows and 16 columns.}
\usage{
qof_meta_org
}
\description{
\tabular{ll}{
 \code{practice_code} \tab {<chr n=8210>} \cr
 \code{practice_name} \tab {<chr n=17148>} \cr
 \code{ccg_code} \tab {<chr n=369>} \cr
 \code{ccg_geography_code} \tab {<chr n=223>} \cr
 \code{ccg_name} \tab {<chr n=377>} \cr
 \code{stp_code} \tab {<chr n=83>} \cr
 \code{stp_name} \tab {<chr n=92>} \cr
 \code{subregion_code} \tab {<chr n=23>} \cr
 \code{subregion_geography_code} \tab {<chr n=22>} \cr
 \code{subregion_name} \tab {<chr n=35>} \cr
 \code{region_code} \tab {<chr n=7>} \cr
 \code{region_geography_code} \tab {<chr n=7>} \cr
 \code{region_name} \tab {<chr n=12>} \cr
 \code{country} \tab {<chr n=2>} \cr
 \code{revised_maximum_points} \tab {<num>} \cr
 \code{qof_period} \tab {<chr n=7>} \cr
}
}
\seealso{
Other qof_meta: \code{\link{qof_datasets}},
  \code{\link{qof_meta_ind}}, \code{\link{qof_meta}}
}
\concept{qof_meta}
\keyword{datasets}
