\name{qofvariation.ncc-package}
\alias{qofvariation.ncc-package}
\alias{qofvariation.ncc}
\docType{package}
\title{
\packageTitle{qofvariation.ncc}
}
\description{
\packageDescription{qofvariation.ncc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qofvariation.ncc}
\packageIndices{qofvariation.ncc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qofvariation.ncc}

Maintainer: \packageMaintainer{qofvariation.ncc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
