% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{restore_data}
\alias{restore_data}
\title{Restore an object}
\usage{
restore_data(name, dir = c("inst/data-int", "data"))
}
\arguments{
\item{name}{(char) name to use}

\item{dir}{(char) directory to save to.

if \code{dir} is \code{data} (default) then usethis::use_data is used.  if is
\code{inst/data-int} then a call to \code{save} is made to the inst/data-int
dir.}
}
\description{
Load an object with specified name.
}
