% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphoci.R
\name{round_nearest}
\alias{round_nearest}
\title{Round}
\usage{
round_nearest(x, digits = 0)
}
\arguments{
\item{x}{number}

\item{digits}{precision}
}
\value{
x rounded to precision
}
\description{
round to nearest whole number
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{proj_path}},
  \code{\link{proj_root}}, \code{\link{setnames.clean}},
  \code{\link{status}}, \code{\link{sum.rmna}},
  \code{\link{transpose}}
}
\concept{Helper routines}
