% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setnames.clean}
\alias{setnames.clean}
\title{Clean data.frame field names}
\usage{
setnames.clean(x)
}
\arguments{
\item{x}{data.frame object}
}
\description{
To clean table/frame names
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{proj_path}},
  \code{\link{proj_root}}, \code{\link{round_nearest}},
  \code{\link{status}}, \code{\link{sum.rmna}},
  \code{\link{transpose}}
}
\concept{Helper routines}
