% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcci.R
\name{stat_test}
\alias{stat_test}
\title{Statistic test routines}
\description{
Statistic test routines
}
\seealso{
Other confidence interval routines: \code{\link{aphoci_gen}},
  \code{\link{aphoci}}, \code{\link{calcci}},
  \code{\link{testci}}, \code{\link{testspc}}

Other statistic test routines: \code{\link{poisson_byars.test}},
  \code{\link{prop_wilson.test}}
}
