% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{store_data}
\alias{store_data}
\title{Store an object}
\usage{
store_data(obj, name, dir = c("inst/data-int", "data"))
}
\arguments{
\item{obj}{(object) object to save}

\item{name}{(char) name to use}

\item{dir}{(char) directory to save to.

if \code{dir} is \code{data} (default) then usethis::use_data is used.  if is
\code{inst/data-int} then a call to \code{save} is made to the inst/data-int
dir.}
}
\description{
Save an object with specified name.
}
