% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum.rmna}
\alias{sum.rmna}
\title{Sum ignoring NAs}
\usage{
\method{sum}{rmna}(x)
}
\arguments{
\item{x}{vector object}
}
\description{
For use in e.g. dcast to ignore NAs
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{transpose}}
}
\concept{Helper routines}
