% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_main}
\alias{test_main}
\title{Test the main routine}
\usage{
test_main(bProcessRaw = TRUE)
}
\arguments{
\item{bProcessRaw}{if FALSE just load.  If TRUE process and then load.}
}
\description{
Iterate over known instances.
goes through 1617 and 1516 both processing raw and loading processed dataset.
}
