% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{testci}
\alias{testci}
\title{Family of testci routines}
\description{
Container for testci routines
}
\seealso{
Other testci routines: \code{\link{testci_gen}}

Other confidence interval routines: \code{\link{aphoci_gen}},
  \code{\link{aphoci}}, \code{\link{calcci}},
  \code{\link{stat_test}}, \code{\link{testspc}}
}
\concept{confidence interval routines}
\concept{testci routines}
