% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{testci_gen}
\alias{testci_gen}
\alias{testci_hilo}
\alias{testci_sense}
\alias{testci_hilo_s}
\alias{testci_sense_s}
\title{Test if two confidence intervals overlap.}
\usage{
testci_gen(ci.ref, ci.var, bAsString = FALSE,
  bSenseHigherisBetter = NA, return_type = c("minimal", "data.frame",
  "data.table"), verbose = FALSE)

testci_hilo(ci.ref, ci.var)

testci_sense(ci.ref, ci.var, bSenseHigherisBetter = NA)

testci_hilo_s(ci.ref, ci.var)

testci_sense_s(ci.ref, ci.var, bSenseHigherisBetter = NA)
}
\arguments{
\item{ci.ref}{Reference point estimate or value range (vector(2))}

\item{ci.var}{Variable point estimate or value range (vector(2))}

\item{bAsString}{bool to specify string over numeric}

\item{bSenseHigherisBetter}{bool to determine any sense checking.  Default is
NA - no sense check.}

\item{return_type}{Specify list of vectors (default) or data.frame or
data.table.}

\item{verbose}{set to TRUE to print some debug output.}
}
\value{
Value indicating relative position of variable to reference value.
Numeric by default, string if \code{bAsString == TRUE}

If the values within the range specified by \code{ci.var} are all lower
than the values within the range specified by \code{ci.ref} then \code{var}
is considered 'signficantly lower' than \code{ref} and the return value of
\code{testci_hilo} is \code{-1}.

Conversely if the range of values specified by \code{ci.var} area all
higher then \code{var} is considered 'significantly HIGHER' than \code{ref}
and the return value is \code{1}.

if the range of values specified by \code{ci.ref} and \code{ci.var} overlap
then \code{ref} and \code{var} are considered 'statistically SIMILAR' and
the return value is \code{0}.

If there is no test done then the return value is \code{NA}.
\itemize{
\item +2 - Better
\item +1 - Higher
\item 0 - Similar
\item -1 - Lower
\item -2 - Worse
\item NA - Not tested
}

Integer indicating relative position of variable to reference value.
\itemize{
\item +2 - Better
\item 0 - Similar
\item -2 - Worse
\item NA - Not tested
}
}
\description{
Generic Version.
}
\section{Functions}{
\itemize{
\item \code{testci_hilo}: Return simple hilo comparison

\item \code{testci_sense}: Apply sense (Better / Worse) to testci comparison

\item \code{testci_hilo_s}: Return a string to denote higher/lower/similar/not tested

\item \code{testci_sense_s}: Return a string to denote better/worse/similar/not tested
}}

\note{
If strings are returned they are from the set {"Worse", "Lower",
"Similar", "Higher", "Better", "Not tested"}.
\itemize{
\item if bHigherisBetter == NA (default) then 1 - Higher, -1 - Lower
\item if bHigherisBetter == TRUE or FALSE then sense of value is considered and 2 -
Better, -2 - Worse.
}
}
\seealso{
Other testci routines: \code{\link{testci}}

Other generic routines: \code{\link{testspc_gen}}

Other returns hilo: \code{\link{testspc_gen}}

Other returns sense: \code{\link{testspc_gen}}

Other returns numeric: \code{\link{testspc_gen}}

Other returns string: \code{\link{testspc_gen}}

Other testci routines: \code{\link{testci}}

Other returns hilo: \code{\link{testspc_gen}}

Other returns numeric: \code{\link{testspc_gen}}

Other testci routines: \code{\link{testci}}

Other returns sense: \code{\link{testspc_gen}}

Other returns numeric: \code{\link{testspc_gen}}

Other testci routines: \code{\link{testci}}

Other returns hilo: \code{\link{testspc_gen}}

Other returns string: \code{\link{testspc_gen}}

Other testci routines: \code{\link{testci}}

Other returns sense: \code{\link{testspc_gen}}

Other returns string: \code{\link{testspc_gen}}
}
\concept{generic routines}
\concept{instances of generic routines}
\concept{returns hilo}
\concept{returns numeric}
\concept{returns sense}
\concept{returns string}
\concept{testci routines}
