% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{testspc}
\alias{testspc}
\title{Family of testspc routines}
\description{
Container for testspc routines
}
\seealso{
Other testspc routines: \code{\link{testspc_gen}}

Other confidence interval routines: \code{\link{aphoci_gen}},
  \code{\link{aphoci}}, \code{\link{calcci}},
  \code{\link{stat_test}}, \code{\link{testci}}
}
