% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{testspc_gen}
\alias{testspc_gen}
\alias{testspc_hilo}
\alias{testspc_hilo_s}
\title{Test variation with SPC methods}
\usage{
testspc_gen(value.var, value.ref, denominator.var, multiplier = 1,
  ci_type = "poisson", sd = 3, bAsString = FALSE,
  return_type = c("minimal", "data.frame", "data.table"))

testspc_hilo(value.var, value.ref, denominator.var, multiplier = 1,
  ci_type = "poisson", sd = 3, return_type = c("minimal", "data.frame",
  "data.table")[1])

testspc_hilo_s(value.var, value.ref, denominator.var, multiplier = 1,
  ci_type = "poisson", sd = 3, return_type = c("minimal", "data.frame",
  "data.table")[1])
}
\arguments{
\item{value.var}{variable value}

\item{value.ref}{reference value}

\item{denominator.var}{variable denominator}

\item{multiplier}{Multiplier.  Optional - default 1.}

\item{ci_type}{Type of confidence interval.  Optional - default "poisson"
Possible values {"poisson", "proportion", "poisson_byars", "prop_wilson",
"poisson_native", "proportion_native", "binomial_native"}. Also byars and
wilson can be used (aliased to poisson_byars and prop_wilson respectively).}

\item{sd}{sds to consider for limit.  2 corresponds to level 95.44997\%, 3 to
99.73002\%.}

\item{bAsString}{bool to specify string over numeric}

\item{return_type}{Specify list of vectors (default) or data.frame or
data.table.}
}
\value{
c(-1, 0, 1)

c("Lower", "Similar", "Higher")
}
\description{
special cause variation detection using SPC methods
Variable value is compared to control limits around a reference value
}
\section{Functions}{
\itemize{
\item \code{testspc_hilo}: Return simple hilo comparison

\item \code{testspc_hilo_s}: Return a string to denote higher/lower/similar/not tested
}}

\seealso{
Other testspc routines: \code{\link{testspc}}

Other generic routines: \code{\link{testci_gen}}

Other returns hilo: \code{\link{testci_gen}}

Other returns sense: \code{\link{testci_gen}}

Other returns numeric: \code{\link{testci_gen}}

Other returns string: \code{\link{testci_gen}}
}
