% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testci.R
\name{transpose}
\alias{transpose}
\title{Tranpose a list of vectors}
\usage{
transpose(l, method = c("native", "data.table", "purrr"))
}
\arguments{
\item{l}{list of length m of vectors length n}

\item{method}{method to use.  Currently ignored.  Default "native".}
}
\value{
list of length n of vectors length m
}
\description{
Takes a list of length m of vectors length n and returns a list of length n
of vectors length m
}
\details{
Native transpose of list of vectors.  Was a wrapper around transpose
routines.  data.table or purrr.
}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{status}},
  \code{\link{sum.rmna}}
}
\concept{Helper routines}
