% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{usethis__use_data}
\alias{usethis__use_data}
\title{Programmatic use of use_data}
\usage{
usethis__use_data(obj, name)
}
\arguments{
\item{obj}{(object) object to save}

\item{name}{(char) name to use}
}
\description{
Call use_data to save object with specified name.
}
