% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qofvariation.ncc-package.R
\docType{data}
\name{verbosity_levels}
\alias{verbosity_levels}
\alias{verbosity_level}
\alias{verbosity.showatlevel}
\title{possible verbosity levels (global)}
\format{An object of class \code{integer} of length 3.}
\usage{
verbosity_levels

verbosity_level

verbosity.showatlevel(this_msg_category)
}
\arguments{
\item{this_msg_category}{}
}
\description{
0 - suppress (default)
1 - information
2 - chatty

0 - suppress (default)
1 - information
2 - chatty

Chatty by default.  Can make less chatty by DECREASING verbosity level
Conversely, INCREASE verbosity with INCREASING values ...
}
\details{
... warnings and errors always show
... switch messages off with 0
... useful stuff for 1
... details at level 2

check level is greater than "suppress" yet lower than "code classification"
}
\examples{
verbosity_level <- verbosity_levels["chatty"]
names(verbosity_levels) \%>\% sapply(verbosity.showatlevel)
names(verbosity_levels) \%>\% sapply(function(x) {
  msg <- paste("global level: ", names(verbosity_level))
  if (verbosity.showatlevel(x)) msg <- paste(msg, "This is a[n]", x, "message")
  msg
})

if (verbosity.showatlevel("info")) cat("Hello World!\\n")

verbosity_level <- verbosity_levels["info"]
names(verbosity_levels) \%>\% sapply(verbosity.showatlevel)

if (verbosity.showatlevel("info")) cat("Hello World!\\n")

verbosity_level <- verbosity_levels["suppress"]
names(verbosity_levels) \%>\% sapply(verbosity.showatlevel)

if (verbosity.showatlevel("info")) cat("Hello World!\\n")

}
\keyword{datasets}
