% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{f__extract__load_raw}
\alias{f__extract__load_raw}
\title{load raw QOF data}
\usage{
f__extract__load_raw(qof_root = c("qof-1617", "qof-1516", "qof-1415",
  "qof-1314")[4], bSaveData = FALSE)
}
\arguments{
\item{qof_root}{Directory root for loading and saving any processed data.  Of the form
\code{qof-YYZZ}}

\item{bSaveData}{(boolean) Flag to determine whether to save datasets as
\code{devtools::use_data}}
}
\value{
a list of lists with named items
\describe{
    \item{reference}{
        \itemize{\item{orgref}\item{indmap}}
    }
    \item{data}{
        \itemize{\item{prev}\item{ind}}
    }
}
}
\description{
put in an R list for later analysis
}
\note{
HF double counted - register_description has HF and HF with LVD
  - keep just HF
}
\seealso{
Other Internal routines: \code{\link{f__91__amend_data__add_subtotals}},
  \code{\link{f__91__amend_orgref__ccg_groups}},
  \code{\link{f__91__compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{f__91__measures_ind}},
  \code{\link{f__91__measures_prev}},
  \code{\link{f__91__measures}},
  \code{\link{f__91__preprocess}},
  \code{\link{f__91__save_reference}},
  \code{\link{families}}

Other Load routines: \code{\link{f__91__load__reference_measures_compare}},
  \code{\link{f__91__load_compare}},
  \code{\link{f__91__load_data}},
  \code{\link{f__91__load_measures}},
  \code{\link{f__91__load_reference}},
  \code{\link{families}}

Other Extract routines: \code{\link{l_readxl_wb}},
  \code{\link{l_readxl_ws_ind}},
  \code{\link{l_readxl_ws_prev}}
}
\concept{Extract routines}
\concept{Internal routines}
\concept{Load routines}
