% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_data_ind}
\alias{qof_data_ind}
\title{QOF data for indicators}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4817811 rows and 8 columns.}
\usage{
qof_data_ind
}
\description{
\tabular{ll}{
 \code{practice_code} \tab {<chr n=8209>} \cr
 \code{indicator_code} \tab {<chr n=255>} \cr
 \code{ACHIEVED_POINTS} \tab {<num>} \cr
 \code{DENOMINATOR} \tab {<num>} \cr
 \code{EXCEPTIONS} \tab {<num>} \cr
 \code{NUMERATOR} \tab {<num>} \cr
 \code{REGISTER} \tab {<num>} \cr
 \code{qof_period} \tab {<chr n=7>} \cr
}
}
\details{
\describe{
  \item{\code{measure}}{
  One of
  \code{ACHIEVED_POINTS}
  , \code{DENOMINATOR}
  , \code{EXCEPTIONS}
  , \code{NUMERATOR}
  , \code{REGISTER}
  }
}
}
\seealso{
Other qof_data: \code{\link{qof_data_prev}},
  \code{\link{qof_datasets}}, \code{\link{qof_data}}
}
\concept{qof_data}
\keyword{datasets}
