% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_data_prev}
\alias{qof_data_prev}
\alias{qof_1718_data_prev}
\alias{qof_1617_data_prev}
\alias{qof_1516_data_prev}
\title{QOF data for prevalence}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1148773 rows and 6 columns.}
\usage{
qof_data_prev

qof_1718_data_prev

qof_1617_data_prev

qof_1516_data_prev
}
\description{
\tabular{ll}{
 \code{practice_code} \tab {<chr n=8210>} \cr
 \code{indicator_group_code} \tab {<chr n=33>} \cr
 \code{register} \tab {<num>} \cr
 \code{patient_list_type} \tab {<chr n=7>} \cr
 \code{patient_list_size} \tab {<num>} \cr
 \code{qof_period} \tab {<chr n=7>} \cr
}
}
\seealso{
Other qof_data: \code{\link{qof_data_ind}},
  \code{\link{qof_datasets}}, \code{\link{qof_data}}
}
\concept{qof_data}
\keyword{datasets}
