% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_meta_ind}
\alias{qof_meta_ind}
\alias{qof_1617_meta_ind}
\alias{qof_1516_meta_ind}
\alias{qof_1415_meta_ind}
\alias{qof_1314_meta_ind}
\alias{qof_1213_meta_ind}
\title{QOF reference data for organisations}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 77 rows and 8 columns.}
\usage{
qof_1617_meta_ind

qof_1516_meta_ind

qof_1415_meta_ind

qof_1314_meta_ind

qof_1213_meta_ind
}
\description{
\preformatted{
$ indicator_code             : chr
$ indicator_description      : chr
$ indicator_point_value      : int
$ indicator_group_code       : chr
$ indicator_group_description: chr
$ domain_code                : chr
$ domain_description         : chr
$ patient_list_type          : chr
}
}
\details{
\describe{
  \item{\code{patient_list_type}}{
  One of
  \code{TOTAL}
  , \code{15OV}, \code{16OV}, \code{17OV}, \code{18OV}
  , \code{45OV}, \code{50OV}
  , \code{25_64_F}
  , \code{30_74}
  , \code{54UN_F}
  }
}
}
\seealso{
Other qof_meta: \code{\link{qof_datasets}},
  \code{\link{qof_meta_org}}, \code{\link{qof_meta}}
}
\concept{qof_meta}
\keyword{datasets}
