% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qof_meta_org}
\alias{qof_meta_org}
\alias{qof_1617_meta_org}
\alias{qof_1516_meta_org}
\alias{qof_1415_meta_org}
\alias{qof_1314_meta_org}
\alias{qof_1213_meta_org}
\title{QOF reference data for organisations}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 7392 rows and 15 columns.}
\usage{
qof_1617_meta_org

qof_1516_meta_org

qof_1415_meta_org

qof_1314_meta_org

qof_1213_meta_org
}
\description{
\preformatted{
$ practice_code           : chr
$ practice_name           : chr
$ ccg_code                : chr
$ ccg_geography_code      : chr
$ ccg_name                : chr
$ stp_code                : chr
$ stp_name                : chr
$ subregion_code          : chr
$ subregion_geography_code: chr
$ subregion_name          : chr
$ region_code             : chr
$ region_geography_code   : chr
$ region_name             : chr
$ country                 : chr
$ revised_maximum_points  : int
}
}
\seealso{
Other qof_meta: \code{\link{qof_datasets}},
  \code{\link{qof_meta_ind}}, \code{\link{qof_meta}}
}
\concept{qof_meta}
\keyword{datasets}
