% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qofvariation.ncc-package.R
\docType{package}
\name{qofvariation.ncc}
\alias{qofvariation.ncc}
\alias{qofvariation.ncc-package}
\title{qofvariation.ncc.}
\description{
Process QOF for visualisation (interactive?)
}
\details{
Require certain disease areas

- practice level prevalence and stat. sig. relative to England
- CDG level achievement (?treatment) and stat. sig. relative to England.
}
