% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg_91_qof.R
\name{status}
\alias{status}
\title{Progress within a pipe}
\usage{
status(x, ...)
}
\arguments{
\item{x}{data object}

\item{...}{message elements.  Passed to cat().}
}
\description{
Show a message then pass on the object invisibly.
}
\note{
\code{"\\n"} added to message.
}
\examples{
\dontrun{
    data \%>\% status("Processing ...") \%>\% some_long_calc()
}

}
\seealso{
Other Helper routines: \code{\link{families}},
  \code{\link{is_installed}}, \code{\link{paste_paths}},
  \code{\link{proj_root}}, \code{\link{round_nearest}},
  \code{\link{setnames.clean}}, \code{\link{sum.rmna}},
  \code{\link{transpose}}
}
